#42 feat(waiting): 기본 페이지 위치 변경
